# virt-exportserver: replace gzip with pgzip