<!--  Thanks for sending a pull request!  Here are some tips for you:
1. Consider creating this PR as draft: https://github.com/kubevirt/kubevirt/blob/main/CONTRIBUTING.md#consider-opening-your-pull-request-as-draft
2. Follow the instructions for writing a release note from k8s: https://git.k8s.io/community/contributors/guide/release-notes.md
-->

### What this PR does
**Before this PR:**
The go standard library `compress/gzip`'s throughput falls short of other common limiting factors such as most modern day commodity hardware due to being single-threaded.

**After this PR:**
This PR replaces `compress/gzip` with [github.com/klauspost/pgzip](https://github.com/klauspost/pgzip), a parallel implementation of gzip that splits the compression process across a number of cores.

pgzip is used in a variety of high-profile OSS projects such as [CockroachDB](https://github.com/cockroachdb/cockroach/pull/88635), [Argo Workflows](https://github.com/argoproj/argo-workflows/pull/5108), [Grafana Loki](https://github.com/grafana/loki/pull/3672), [Podman](https://github.com/containers/podman/pull/22041) and more.

Comparison of various compression softwares has been conducted for the `virt-exportserver`.
The test used a large volume consisting of multiple high and low entropy data including [Silesia Corpus](https://sun.aei.polsl.pl//~sdeor/index.php?page=silesia).

Configuration | CPU Util | RAM Util | Avg Throughput | Time | Compressed Size
-- | -- | -- | -- | -- | --
raw | 1495m | 69Mi | 1081M/s | 02:57 | 187Gi (full)
pgzip 1MiB x2 | 1400m | 71Mi | 95.2M/s | 02:10 | 12.1Gi
pgzip 1MiB x4 | 1491m | 80Mi | 100M/s | 02:03 | 12.1Gi
pgzip 1MiB x8 | 1660m | 95Mi | 102M/s | 02:00 | 12.1Gi
pgzip 4MiB x2 | 1633m | 117Mi | 100M/s | 02:03 | 12.0Gi
pgzip 4MiB x4 | 1650m | 122Mi | 102M/s | 02:00 | 12.0Gi
pgzip 4MiB x8 | 1700m | 131Mi | 99M/s | 02:04 | 12.0Gi
pgzip 4MiB x8 (level=4) | 1630m | 128Mi | 97.6M/s | 02:06 | 12.0Gi
compress/gzip | 995m | 34Mi | 19.4M/s | 10:31 | 12.0Gi
GNU gzip | 993m | 32Mi | 21.4M/s | 10:22 | 12.0Gi
pigz | 3375m | 38Mi | 38.7M/s | 05:18 | 12.0Gi

In conclusion, I've decided to use pgzip with default configuration (i.e., 1MiB block-size, `runtime.GOMAXPROCS(0)` blocks) for this usecase as drop-in replacement for gzip.

<!-- (optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close the issue(s) when PR gets merged)*: -->

### Why we need it and why it was done in this way
The following tradeoffs were made:

The following alternatives were considered:

Links to places where the discussion took place: <!-- optional: slack, other GH issue, mailinglist, ... -->

### Special notes for your reviewer

<!-- optional -->

### Checklist

This checklist is not enforcing, but it's a reminder of items that could be relevant to every PR.
Approvers are expected to review this list.

- [ ] Design: A [design document](https://github.com/kubevirt/community/tree/main/design-proposals) was considered and is present (link) or not required
- [ ] PR: The PR description is expressive enough and will help future contributors
- [ ] Code: [Write code that humans can understand](https://en.wikiquote.org/wiki/Martin_Fowler#code-for-humans) and [Keep it simple](https://en.wikipedia.org/wiki/KISS_principle)
- [ ] Refactor: You have [left the code cleaner than you found it (Boy Scout Rule)](https://learning.oreilly.com/library/view/97-things-every/9780596809515/ch08.html)
- [ ] Upgrade: Impact of this change on upgrade flows was considered and addressed if required
- [ ] Testing: New code requires [new unit tests](https://github.com/kubevirt/kubevirt/blob/main/docs/reviewer-guide.md#when-is-a-pr-good-enough). New features and bug fixes require at least on e2e test
- [ ] Documentation: A [user-guide update](https://github.com/kubevirt/user-guide/) was considered and is present (link) or not required. You want a user-guide update if it's a user facing feature / API change.
- [ ] Community: Announcement to [kubevirt-dev](https://groups.google.com/g/kubevirt-dev/) was considered

### Release note
<!--  Write your release note:
1. Enter your extended release note in the below block. If the PR requires additional action from users switching to the new release, include the string "action required".
2. If no release note is required, just write "NONE".
-->
```release-note
NONE
```